+++ src/Expensify.js
@@ -5,7 +5,7 @@ import SignInPage from './page/SignInPage';
 import HomePage from './page/HomePage/HomePage';
 import Ion from './lib/Ion';
 import * as ActiveClientManager from './lib/ActiveClientManager';
-import {verifyAuthToken} from './lib/actions/ActionsSession';
+import verifyAuthToken from './lib/actions/ActionReauthenticate';
 import IONKEYS from './IONKEYS';
 import WithIon from './components/WithIon';
 import {
